requirements.txt
# Core dependencies
numpy>=1.20.0
torch>=1.10.0
python-dotenv>=0.19.0
huggingface_hub>=0.10.0

# Search results retrieval
requests>=2.25.1
beautifulsoup4>=4.9.3
newspaper3k>=0.2.8

# PDF support (optional but recommended)
PyPDF2>=2.0.0
pdfplumber>=0.7.0

# Embedding generation
voyageai>=0.1.0

# Query generation
openai>=1.0.0

# Additional dependencies that might be needed
transformers>=4.15.0  # For Hugging Face models
